fix missed renames
