+++ frontend/src/features/collections/index.ts
@@ -2,7 +2,7 @@ import("./collections-add");
 import("./collections-grid");
 import("./collection-items-dialog");
 import("./collection-edit-dialog");
-import("./collection-metadata-dialog");
+import("./collection-create-dialog");
 import("./collection-replay-dialog");
 import("./collection-workflow-list");
 import("./select-collection-access");
