+++ frontend/src/pages/org/collections-list.ts
@@ -14,8 +14,8 @@ import type { SelectNewDialogEvent } from ".";
 import { BtrixElement } from "@/classes/BtrixElement";
 import type { PageChangeEvent } from "@/components/ui/pagination";
 import { ClipboardController } from "@/controllers/clipboard";
+import type { CollectionSavedEvent } from "@/features/collections/collection-create-dialog";
 import { type EditDialogTab } from "@/features/collections/collection-edit-dialog";
-import type { CollectionSavedEvent } from "@/features/collections/collection-metadata-dialog";
 import { SelectCollectionAccess } from "@/features/collections/select-collection-access";
 import { emptyMessage } from "@/layouts/emptyMessage";
 import { pageHeader } from "@/layouts/pageHeader";
